add sanitize options for toXml function. (& -> &amp;, etc.)
