+++ lib/json2xml.js
@@ -1,7 +1,21 @@
-module.exports = function toXml(json, xml) {
+var san = require('./sanitize.js')
+
+module.exports = function toXml(json, xml, _options) {
     var xml = xml || '';
+    if (_options === undefined) {
+        _options = xml;
+        xml = '';
+    }
+    
     if (json instanceof Buffer) {
         json = json.toString();
+    }
+
+    options = {
+        sanitize: false
+    };
+    for (var opt in _options) {
+        options[opt] = _options[opt];
     }
 
     var obj = null;
@@ -17,7 +31,7 @@ module.exports = function toXml(json, xml) {
 
     var keys = Object.keys(obj);
     var len = keys.length;
-
+    
     // First pass, extract strings only
     for (var i = 0; i < len; i++) {
         var key = keys[i], value = obj[key], isArray = Array.isArray(value);
@@ -30,6 +44,9 @@ module.exports = function toXml(json, xml) {
 						xml += subVal;
 					} else {
 						xml = xml.replace(/>$/, '');
+                        if (options.sanitize) {
+                            subVal = san.sanitize(subVal);
+                        }
 						xml += ' ' + key + '="' + subVal + '">';
 					}
 				}
@@ -49,13 +66,13 @@ module.exports = function toXml(json, xml) {
 
 				if (typeof(elem) == 'object') {
 					xml += '<' + key + '>';
-					xml = toXml(elem, xml);
+					xml = toXml(elem, xml, options);
 					xml += '</' + key + '>';
 				}
             }
         } else if (typeof(obj[key]) == 'object') {
             xml += '<' + key + '>';
-            xml = toXml(obj[key], xml);
+            xml = toXml(obj[key], xml, options);
             xml += '</' + key + '>';
         }
     }
