+++ lib/xml2json.js
@@ -1,4 +1,5 @@
 var expat = require('node-expat');
+var san = require('./sanitize.js')
 
 // This object will hold the final result.
 var obj = {};
@@ -58,7 +59,7 @@ function text(data) {
     }
 
     if (options.sanitize) {
-        data = sanitize(data);
+        data = san.sanitize(data);
     }
 
     currentObject['$t'] = coerce((currentObject['$t'] || '') + data);
@@ -107,41 +108,6 @@ function coerce(value) {
     return value;
 }
 
-
-/**
- * Simple sanitization. It is not intended to sanitize
- * malicious element values.
- *
- * character | escaped
- *      <       &lt;
- *      >       &gt;
- *      (       &#40;
- *      )       &#41;
- *      #       &#35;
- *      &       &amp;
- *      "       &quot;
- *      '       &apos;
- */
-var chars =  {  '<': '&lt;',
-                '>': '&gt;',
-                '(': '&#40;',
-                ')': '&#41;',
-                '#': '&#35;',
-                '&': '&amp;',
-                '"': '&quot;',
-                "'": '&apos;' };
-
-function sanitize(value) {
-    if (typeof value !== 'string') {
-        return value;
-    }
-
-    Object.keys(chars).forEach(function(key) {
-        value = value.replace(key, chars[key]);
-    });
-
-    return value;
-}
 
 /**
  * Parses xml to json using node-expat.
