+++ test/fixtures/xmlsanitize.xml
@@ -1,0 +1,1 @@
+<e><a b="Smith &amp; Son"></a></e>
